set aws S3  (#74)
* feat update productdetailsapi

* feat : add isFavorite for details api

* feat : doing set awsS3

* feat : set awss3

+++ src/main/java/com/iko/iko/service/product/GetAllProductService.java
@@ -56,7 +56,7 @@ public class GetAllProductService {
             }
             ProductDetailsResponse.MainProductForResponse checkData
                     =new ProductDetailsResponse.MainProductForResponse(
-                            isFavorite,
+                    (int)(long)isFavorite,
                     tmp.getProductId(),
                     tmp.getSeries(),
                     gList,
